LayoutTests/http/tests/performance/paint-timing/performance-paint-timing-fcp-after-visually-non-empty-for-num-chars-expected.txt

PASS paintEntry.startTime > timestamps[expectedIterations] is true
PASS paintEntry.startTime < timestamps[expectedIterations + 2] is true
